Make property presence globally optional
Also remove incorrect "targets_optional" for midi-key-in
and websocket - what was actually meant is for midi-key-in
is some properties are optional.  Not sure why websocket ever
had it.

Need to come back to this and mark individual properties as
optional, enabled only if specifically mentioned - or define
a single property as default which can be assumed...

Start of close encounters restructure that uses it, too