Make service lookup by section & type, and dynamic
Service lookups were previously done at setup() timem which caused
two problems:

1) Creation became order-dependent, because a subgraph might be
created and configured before the service in its parent that it
depends on (actual, blocking JSON create)

2) There was a nasty pointer dependency on the service elements,
which could be replaced dynamically (potential)

Lookup by type means it doesn't matter what the ID is, however
it is also currently a linear search, now done on every tick or
event.  Could be optimised with a local service map on the graph
if it's a problem.