Memory takes average, add drag to pendulum
Trying to reduce the delay causing overswing of the pendulum,
but only partially successful - the bigger factor is that the two
interconnected memories introduce a 2 tick delay in the calculation,
which means RV is still increasing as it passes angle=0 - this makes
the pendulum swing increase each time.

Added drag factor to counteract this...  Not clear if a real solution is
possible!

bug:comment:1559